+++ packages/browser-extension/src/modules/service/utils/sort.ts
@@ -1,16 +1,27 @@
 import { DbWorkItem } from "../../db/db";
+import { MetadataMap } from "../emitters/metadata-manager";
 
-export function sortByState(a: DbWorkItem, b: DbWorkItem): number {
-  return getStatePriority(a.state) - getStatePriority(b.state);
+export function sortByState(metadata: MetadataMap, a: DbWorkItem, b: DbWorkItem): number {
+  const categoryA = metadata.get(a.workItemType)?.states.get(a.state)?.category;
+  const categoryB = metadata.get(b.workItemType)?.states.get(b.state)?.category;
+  return getCategoryPriority(categoryA) - getCategoryPriority(categoryB);
 }
 
-function getStatePriority(state: string): number {
-  switch (state) {
-    case "Completed":
-      return 1;
-    case "Cut":
-      return 2;
-    default:
+// Ref: https://docs.microsoft.com/en-us/azure/devops/boards/work-items/workflow-and-state-categories
+function getCategoryPriority(category?: string): number {
+  switch (category) {
+    case "Proposed":
       return 0;
+    case "InProgress":
+      return 1;
+    case "Resolved":
+      return 8;
+    case "Completed":
+      return 9;
+    case "Removed":
+      return 10;
+
+    default:
+      return 5;
   }
 }